:tada: feat(reaction-role): minor refactor
